Renamed components for a more cohesive naming convention
