+++ src/App.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 import { AugmentedFormula } from "./FormulaTree";
 import { FormulizeConfig } from "./api";
 import { computationStore } from "./api/computation";
-import Header from "./components/header.tsx";
+import Header from "./components/navigation.tsx";
 import APIPage from "./pages/api/index.tsx";
 import EditorPage from "./pages/editor/EditorPage";
 import { formulaStore } from "./store";
